Style improvements: Removed redundant initializers, std::move where possible, use .contains on containers, a few other changes. No functional differences.
[notest]

+++ src/AST/NodeTypes.h
@@ -6,8 +6,10 @@
 
 #pragma once
 
+#include <cstdint>
+
 namespace AST {
-enum class NodeType : int {
+enum class NodeType : uint8_t {
 	Program = 0,
 	ArrayAssignment,
 	ArrayIndex,
